requirements.txt
adafruit-circuitpython-neopixel
adafruit-circuitpython-neopixel-spi
